Rename buff and action files
