+++ engine/action/action_callback.cpp
@@ -4,7 +4,7 @@
 // ==========================================================================
 
 #include "action_callback.hpp"
-#include "sc_action.hpp"
+#include "action.hpp"
 #include "player/player.hpp"
 
 action_callback_t::action_callback_t( player_t* l, bool ap, bool asp ) :
