+++ engine/action/action_state.cpp
@@ -3,8 +3,8 @@
 // Send questions to natehieter@gmail.com
 // ==========================================================================
 
-#include "action/sc_action_state.hpp"
-#include "action/sc_action.hpp"
+#include "action/action_state.hpp"
+#include "action/action.hpp"
 #include "player/player.hpp"
 #include "sim/sim.hpp"
 #include <sstream>
