+++ engine/action/dbc_proc_callback.cpp
@@ -7,12 +7,12 @@
 
 #include <cassert>
 
-#include "buff/sc_buff.hpp"
+#include "buff/buff.hpp"
 #include "item/item.hpp"
 #include "item/special_effect.hpp"
 #include "player/player.hpp"
-#include "sc_action.hpp"
-#include "sc_action_state.hpp"
+#include "action.hpp"
+#include "action_state.hpp"
 #include "action_callback.hpp"
 #include "sim/event.hpp"
 #include "sim/real_ppm.hpp"
