+++ engine/player/action_priority_list.cpp
@@ -5,7 +5,7 @@
 
 #include "action_priority_list.hpp"
 
-#include "action/sc_action.hpp"
+#include "action/action.hpp"
 #include "dbc/dbc.hpp"
 #include "player/player.hpp"
 #include "util/util.hpp"
