+++ engine/player/consumable.cpp
@@ -8,9 +8,9 @@
 #include "dbc/dbc.hpp"
 #include "sc_enums.hpp"
 #include "player.hpp"
-#include "action/sc_action.hpp"
+#include "action/action.hpp"
 #include "action/heal.hpp"
-#include "buff/sc_buff.hpp"
+#include "buff/buff.hpp"
 #include "sim/cooldown.hpp"
 #include "sim/expressions.hpp"
 #include "sim/sim.hpp"
