+++ engine/player/player.cpp
@@ -11,13 +11,13 @@
 #include "action/dot.hpp"
 #include "action/heal.hpp"
 #include "action/residual_action.hpp"
-#include "action/sc_action.hpp"
-#include "action/sc_action_state.hpp"
+#include "action/action.hpp"
+#include "action/action_state.hpp"
 #include "action/sequence.hpp"
 #include "action/snapshot_stats.hpp"
 #include "action/spell.hpp"
 #include "action/variable.hpp"
-#include "buff/sc_buff.hpp"
+#include "buff/buff.hpp"
 #include "dbc/active_spells.hpp"
 #include "dbc/azerite.hpp"
 #include "dbc/dbc.hpp"
