+++ engine/player/soulbinds.cpp
@@ -5,7 +5,7 @@
 #include "soulbinds.hpp"
 
 #include "action/dot.hpp"
-#include "action/sc_action_state.hpp"
+#include "action/action_state.hpp"
 #include "item/item.hpp"
 #include "player/actor_target_data.hpp"
 #include "player/covenant.hpp"
