+++ engine/sim/expressions.cpp
@@ -4,7 +4,7 @@
 // ==========================================================================
 
 #include "expressions.hpp"
-#include "action/sc_action.hpp"
+#include "action/action.hpp"
 #include "player/player.hpp"
 #include "sim/sim.hpp"
 #include <atomic>
