+++ source_files/QT_engine.pri
@@ -2,20 +2,20 @@
 # To change the list of source files run synchronize.py
 
 HEADERS += engine/action/absorb.hpp
+HEADERS += engine/action/action.hpp
 HEADERS += engine/action/action_callback.hpp
+HEADERS += engine/action/action_state.hpp
 HEADERS += engine/action/attack.hpp
 HEADERS += engine/action/dbc_proc_callback.hpp
 HEADERS += engine/action/dot.hpp
 HEADERS += engine/action/heal.hpp
 HEADERS += engine/action/residual_action.hpp
-HEADERS += engine/action/sc_action.hpp
-HEADERS += engine/action/sc_action_state.hpp
 HEADERS += engine/action/sequence.hpp
 HEADERS += engine/action/snapshot_stats.hpp
 HEADERS += engine/action/spell.hpp
 HEADERS += engine/action/spell_base.hpp
 HEADERS += engine/action/variable.hpp
-HEADERS += engine/buff/sc_buff.hpp
+HEADERS += engine/buff/buff.hpp
 HEADERS += engine/class_modules/apl/apl_death_knight.hpp
 HEADERS += engine/class_modules/apl/apl_hunter.hpp
 HEADERS += engine/class_modules/apl/apl_monk.hpp
@@ -182,20 +182,20 @@ HEADERS += engine/util/vector_with_callback.hpp
 HEADERS += engine/util/xml.hpp
 
 SOURCES += engine/action/absorb.cpp
+SOURCES += engine/action/action.cpp
 SOURCES += engine/action/action_callback.cpp
+SOURCES += engine/action/action_state.cpp
+SOURCES += engine/action/attack.cpp
 SOURCES += engine/action/dbc_proc_callback.cpp
+SOURCES += engine/action/dot.cpp
 SOURCES += engine/action/heal.cpp
 SOURCES += engine/action/residual_action.cpp
-SOURCES += engine/action/sc_action.cpp
-SOURCES += engine/action/sc_action_state.cpp
-SOURCES += engine/action/sc_attack.cpp
-SOURCES += engine/action/sc_dot.cpp
-SOURCES += engine/action/sc_spell.cpp
-SOURCES += engine/action/sc_stats.cpp
 SOURCES += engine/action/sequence.cpp
 SOURCES += engine/action/snapshot_stats.cpp
+SOURCES += engine/action/spell.cpp
+SOURCES += engine/action/stats.cpp
 SOURCES += engine/action/variable.cpp
-SOURCES += engine/buff/sc_buff.cpp
+SOURCES += engine/buff/buff.cpp
 SOURCES += engine/class_modules/apl/apl_death_knight.cpp
 SOURCES += engine/class_modules/apl/apl_hunter.cpp
 SOURCES += engine/class_modules/apl/apl_monk.cpp
