+++ source_files/VS_engine.props
@@ -6,20 +6,20 @@ To change the list of source files run synchronize.py
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 	<ItemGroup>
 		<ClInclude Include="..\engine\action\absorb.hpp" />
+		<ClInclude Include="..\engine\action\action.hpp" />
 		<ClInclude Include="..\engine\action\action_callback.hpp" />
+		<ClInclude Include="..\engine\action\action_state.hpp" />
 		<ClInclude Include="..\engine\action\attack.hpp" />
 		<ClInclude Include="..\engine\action\dbc_proc_callback.hpp" />
 		<ClInclude Include="..\engine\action\dot.hpp" />
 		<ClInclude Include="..\engine\action\heal.hpp" />
 		<ClInclude Include="..\engine\action\residual_action.hpp" />
-		<ClInclude Include="..\engine\action\sc_action.hpp" />
-		<ClInclude Include="..\engine\action\sc_action_state.hpp" />
 		<ClInclude Include="..\engine\action\sequence.hpp" />
 		<ClInclude Include="..\engine\action\snapshot_stats.hpp" />
 		<ClInclude Include="..\engine\action\spell.hpp" />
 		<ClInclude Include="..\engine\action\spell_base.hpp" />
 		<ClInclude Include="..\engine\action\variable.hpp" />
-		<ClInclude Include="..\engine\buff\sc_buff.hpp" />
+		<ClInclude Include="..\engine\buff\buff.hpp" />
 		<ClInclude Include="..\engine\class_modules\apl\apl_death_knight.hpp" />
 		<ClInclude Include="..\engine\class_modules\apl\apl_hunter.hpp" />
 		<ClInclude Include="..\engine\class_modules\apl\apl_monk.hpp" />
@@ -185,20 +185,20 @@ To change the list of source files run synchronize.py
 		<ClInclude Include="..\engine\util\vector_with_callback.hpp" />
 		<ClInclude Include="..\engine\util\xml.hpp" />
 		<ClCompile Include="..\engine\action\absorb.cpp" />
+		<ClCompile Include="..\engine\action\action.cpp" />
 		<ClCompile Include="..\engine\action\action_callback.cpp" />
+		<ClCompile Include="..\engine\action\action_state.cpp" />
+		<ClCompile Include="..\engine\action\attack.cpp" />
 		<ClCompile Include="..\engine\action\dbc_proc_callback.cpp" />
+		<ClCompile Include="..\engine\action\dot.cpp" />
 		<ClCompile Include="..\engine\action\heal.cpp" />
 		<ClCompile Include="..\engine\action\residual_action.cpp" />
-		<ClCompile Include="..\engine\action\sc_action.cpp" />
-		<ClCompile Include="..\engine\action\sc_action_state.cpp" />
-		<ClCompile Include="..\engine\action\sc_attack.cpp" />
-		<ClCompile Include="..\engine\action\sc_dot.cpp" />
-		<ClCompile Include="..\engine\action\sc_spell.cpp" />
-		<ClCompile Include="..\engine\action\sc_stats.cpp" />
 		<ClCompile Include="..\engine\action\sequence.cpp" />
 		<ClCompile Include="..\engine\action\snapshot_stats.cpp" />
+		<ClCompile Include="..\engine\action\spell.cpp" />
+		<ClCompile Include="..\engine\action\stats.cpp" />
 		<ClCompile Include="..\engine\action\variable.cpp" />
-		<ClCompile Include="..\engine\buff\sc_buff.cpp" />
+		<ClCompile Include="..\engine\buff\buff.cpp" />
 		<ClCompile Include="..\engine\class_modules\apl\apl_death_knight.cpp" />
 		<ClCompile Include="..\engine\class_modules\apl\apl_hunter.cpp" />
 		<ClCompile Include="..\engine\class_modules\apl\apl_monk.cpp" />
