+++ source_files/cmake_engine.txt
@@ -1,19 +1,19 @@
 set(source_files
 action/absorb.hpp
+action/action.hpp
 action/action_callback.hpp
+action/action_state.hpp
 action/attack.hpp
 action/dbc_proc_callback.hpp
 action/dot.hpp
 action/heal.hpp
 action/residual_action.hpp
-action/sc_action.hpp
-action/sc_action_state.hpp
 action/sequence.hpp
 action/snapshot_stats.hpp
 action/spell.hpp
 action/spell_base.hpp
 action/variable.hpp
-buff/sc_buff.hpp
+buff/buff.hpp
 class_modules/apl/apl_death_knight.hpp
 class_modules/apl/apl_hunter.hpp
 class_modules/apl/apl_monk.hpp
@@ -179,20 +179,20 @@ util/util.hpp
 util/vector_with_callback.hpp
 util/xml.hpp
 action/absorb.cpp
+action/action.cpp
 action/action_callback.cpp
+action/action_state.cpp
+action/attack.cpp
 action/dbc_proc_callback.cpp
+action/dot.cpp
 action/heal.cpp
 action/residual_action.cpp
-action/sc_action.cpp
-action/sc_action_state.cpp
-action/sc_attack.cpp
-action/sc_dot.cpp
-action/sc_spell.cpp
-action/sc_stats.cpp
 action/sequence.cpp
 action/snapshot_stats.cpp
+action/spell.cpp
+action/stats.cpp
 action/variable.cpp
-buff/sc_buff.cpp
+buff/buff.cpp
 class_modules/apl/apl_death_knight.cpp
 class_modules/apl/apl_hunter.cpp
 class_modules/apl/apl_monk.cpp
