+++ source_files/engine_make
@@ -3,20 +3,20 @@
 
 SRC += \
     action$(PATHSEP)absorb.cpp \
+    action$(PATHSEP)action.cpp \
     action$(PATHSEP)action_callback.cpp \
+    action$(PATHSEP)action_state.cpp \
+    action$(PATHSEP)attack.cpp \
     action$(PATHSEP)dbc_proc_callback.cpp \
+    action$(PATHSEP)dot.cpp \
     action$(PATHSEP)heal.cpp \
     action$(PATHSEP)residual_action.cpp \
-    action$(PATHSEP)sc_action.cpp \
-    action$(PATHSEP)sc_action_state.cpp \
-    action$(PATHSEP)sc_attack.cpp \
-    action$(PATHSEP)sc_dot.cpp \
-    action$(PATHSEP)sc_spell.cpp \
-    action$(PATHSEP)sc_stats.cpp \
     action$(PATHSEP)sequence.cpp \
     action$(PATHSEP)snapshot_stats.cpp \
+    action$(PATHSEP)spell.cpp \
+    action$(PATHSEP)stats.cpp \
     action$(PATHSEP)variable.cpp \
-    buff$(PATHSEP)sc_buff.cpp \
+    buff$(PATHSEP)buff.cpp \
     class_modules$(PATHSEP)apl$(PATHSEP)apl_death_knight.cpp \
     class_modules$(PATHSEP)apl$(PATHSEP)apl_hunter.cpp \
     class_modules$(PATHSEP)apl$(PATHSEP)apl_monk.cpp \
